requirements.txt
flet>=0.21.0
nicegui>=1.4.0
pywebview>=4.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
